Add interactive stub for reservation
This endpoint pretends that all orgs have
an id quota of 5. No state is saved so that
a variety of edge case requests can be made
against it with predictable results to aid
development of tools that will interact with
this endpoint.

If a request is made that will go over the id
quota or any of the parameters are invalid,
clear error responses are returned.

Blatantly fake IDs are returned, but are the
proper amount according to the amount query
parameter. They also comply with being
sequential or nonsequential. Their structure
is also identical to what the production
service intends to return.