Stop id_quota getting reset on Org updates
Some of the logic around making sure id_quota gets set to zero when the
org is the Secretariat caused all regular orgs to have their id_quota
reset upon any update without a quota provided in that update. Now,
the original quota won't get wiped if a quota is not included.