remove shared `docker/.docker-env`
I'm going to add a `debug` version of the application, where the Node
inspector can be accessed on port `9229`. Using this would require a
new flag to get passed to the `node-dev` process. Switching between the
current `development` and a new `debug` local (docker) configuration
could result in committing different instantiations of the local
environment `docker/.docker-env` to Git. To avoid this, that file is
removed and replaced by an example file.