prevent unverified `jsonschema` package updates
Developers have noticed two tests that fail when they are run in docker
but that pass when they are run locally. These two tests were failing
due to an unexamined (by me, anyway) error in the `jsonschema` package
`Validator`. The tests pass for versions `1.2.x` but fail in `1.3.0`
and `1.4.0`. Assuming that these tests need to pass for the IDR to
maintain its functionality, `package.json` has been updated to specify
`<1.3.0` for `jsonschema`. The tests now pass in docker.

Why were the tests failing? Both tests fail in `src/middleware/
middleware.js/validateCveJsonSchema()` when the state of the CVE being
validated is `PUBLIC`, with the `jsonschema.Validator().validate`
throwing an error we don't catch. The error is `no such schema
https://cve.mitre.org/https://www.first.org/cvss/cvss-v3.1.json?20190610`
which implies we aren't using the `jsonschema.Validator().addSchema`
method properly. It isn't clear if we are using it properly in versions
`<1.3.0` or if we were never using it properly but newer versions of
`jsonschema` catch an error they weren't catching before.

Commentary in the `jsonschema` GitHub repo suggests that previous
versions were failing silently for invalid schemas. It might make more
sense to undo this commit and to investigate whether we were ever
properly using the `addSchema` method, or if we were giving it an
invalid schema but it wasn't noticing. I'm keeping this commit to
document this issue.