Change response to always return new object
This PR starts the beginning of a refactor. Instead of
keeping all errors as constants, we should move to having an error file
for each controller that contains functions for returning the errors
we need for that controller. This way, errors can start becoming dynamic
and the objects are encapsulated so we don't risk sharing objects across
endpoints again.