we will respect our package lock
Copy the package lock file into both production and local containers so
the `npm install` layer respects it. To generate a new lock file the
developer could empty the contents of `package-lock.json` before it
gets copied into the local (dev) container. A better option should be
provided to developers (and can be ... time permitting).