update docker readme for environment configuration
A recent pull request removed the repo's `docker/.docker-env` in favor
of a `docker/.docker-env.example`, without updating `docker/docker-compose.yml`
to look for the new file. This is a safeguard against developers
committing local docker environment changes, since it forces them to
create a `docker/.docker-env` which is included in our `.gitignore`.
However, that pull request did not update `docker/README.md` with an
instruction for users to copy the environment file.