make `src/repositories` file names consistent
The modules located at `src/repositories` used mixed camel case and
what what I'll call big-headed-camel case (camel case that starts with
a capital letter instead of a lowercase letter). I'm not sure if
there's a convention here, but the rest of the `.js` files in the
repository use camel case. The confusion likely comes from a convention
of using big-headed-camel case for models, and that these repositories
are mimicking our data models.