remove unused dependencies
The `husky` package is not in use. It is used to create git hooks,
which are defined in `package.json`. We have none of these definitions
and we plan to use GitHub Actions for that kind of workflow. The
package can be added back if we have an operations or development
workflow requirement that GitHub actions can't meet.

The `standard` package does not connect to the code. It is a dev tool
that could be added to our `package.json` scripts for linting. We are
currently using `eslint` but can add `standard` back in at the request
of any developers would prefer it over `eslint`.

The `axios` package is not used anywhere in the source code. The `uuid`
package is heavily leveraged by the project. It does not belong in the
`devDependencies` section of our `package.json`.

All of our tests pass when these packages are removed.