do not install dev dependencies in production
Testing revealed that the `kleur` package is used, even if it isn't
required or loaded anywhere. The `depcheck` package is technically used
in our `clean` command but can be moved to `devDependencies`.

For now, that is for local setup reasons that would be frustrating to
change, we need to include `node-dev` in our production dependencies.
This might make sense given that we're using the same `Dockerfile` to
deploy to development, staging and production environments, as well as
to a local setup.