Use '99-0002' in CVE ID tests
The ID CVE-1999-0001 used for CVE ID testing has
a Record associated with it that gets inserted along
with the other test fixtures. With the new business
logic, modification of a CVE ID can't happen when a
record exists for the CVE ID. CVE-1999-0002 does NOT
have a record associated with it, and therefore is
a great candidate for replacing 0001 for the time being.